feat(models): update

+++ src/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Boolean, Column, Integer, String
+from sqlalchemy import Boolean, Column, Float, Integer, String
 
 from database import Base
 
@@ -8,7 +8,7 @@ class Payment(Base):
 
     id = Column(Integer, primary_key=True, nullable=False)
     user = Column(Integer, nullable=False)
-    amount = Column(Integer, nullable=False)
+    amount = Column(Float, nullable=False)
     currency = Column(String, nullable=False)
     date = Column(String, nullable=False)
     description = Column(String, nullable=False)
